Insert Assertions
Tasks have to be checked if they are converted correctly from user's
input.

Let's add assertions for Task.java to check if description and isDone is
not empty and not done respectively.
Let's check that each of the tasks are created successfully and return the
string correctly to Cleo.java.

Errors will be thrown if code fails assertions.